Fix PUB/SUB TCP test stall
Turns out this is because the TCP needs time to handshake the
connection, and ZMQ makes no guarantee that a SUB connects before
a PUB sends (and anything sent before the connection is made, is
lost...).

For the time being, a 1 second sleep should do, but a better
solution should be found.